Split setup files into grouped actions
This makes it easier to run specific components individually

+++ setup_symlinks.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+array_contains () {
+    local e match="$1"
+    shift
+    for e; do [[ "$match" =~ "$e" ]] && return 0; done
+    return 1
+}
+
 echo "Symlinking user dirs"
 # must happen before files
 user_dirs=(
@@ -9,6 +16,8 @@ user_dirs=(
 
 for dir in "${user_dirs[@]}"; do
     dest_dir=${dir/"user"/$HOME}
+    # skip if dir is already symlinked
+    if [[ -L "$dest_dir" && -d "$dest_dir" ]]; then continue; fi
     parent_dir=$(echo $dir | sed -E 's/\/[^/]*$//')
     mkdir -p $parent_dir
     ln -sf $(pwd)/$dir $dest_dir
@@ -18,6 +27,8 @@ echo "Symlinking user files"
 user_files=$(find user -type f)
 
 for file in $user_files; do
+    # skip files in the symlinked directories
+    if array_contains $file "${user_dirs[@]}"; then continue; fi
     dest_file=${file/"user"/$HOME}
     dir=$(echo $dest_file | sed -E 's/\/[^/]*$//')
     mkdir -p $dir
